fix codex warning

+++ ecmascript/builtins/builtins_reflect.cpp
@@ -39,7 +39,7 @@ JSTaggedValue BuiltinsReflect::ReflectApply(EcmaRuntimeCallInfo *argv)
 
     // 3. Perform PrepareForTailCall().
     // 4. Return ? Call(target, thisArgument, args).
-    const int32_t argsLength = args->GetLength();
+    const int32_t argsLength = static_cast<int32_t>(args->GetLength());
     JSHandle<JSTaggedValue> undefined = thread->GlobalConstants()->GetHandledUndefined();
     EcmaRuntimeCallInfo *info =
         EcmaInterpreter::NewRuntimeCallInfo(thread, target, thisArgument, undefined, argsLength);
@@ -74,7 +74,7 @@ JSTaggedValue BuiltinsReflect::ReflectConstruct(EcmaRuntimeCallInfo *argv)
     RETURN_EXCEPTION_IF_ABRUPT_COMPLETION(thread);
     JSHandle<TaggedArray> args = JSHandle<TaggedArray>::Cast(argOrAbrupt);
     // 5. Return ? Construct(target, args, newTarget).
-    const int32_t argsLength = args->GetLength();
+    const int32_t argsLength = static_cast<int32_t>(args->GetLength());
     JSHandle<JSTaggedValue> undefined = thread->GlobalConstants()->GetHandledUndefined();
     EcmaRuntimeCallInfo *info = EcmaInterpreter::NewRuntimeCallInfo(thread, target, undefined, newTarget, argsLength);
     RETURN_EXCEPTION_IF_ABRUPT_COMPLETION(thread);
